+++ web/src/app/(customer)/travel-plans/_components/Itinerary.tsx
@@ -1,0 +1,10 @@
+export const Itinerary = () => {
+  return (
+    <div
+      className="text-[50px] font-bold"
+      style={{ fontFamily: "Dancing Script" }}
+    >
+      Itinerary
+    </div>
+  );
+};
